recursive currency calculation
