+++ Content.Shared/_CP14/Currency/CP14CurrencyExaminableComponent.cs
@@ -1,0 +1,10 @@
+namespace Content.Shared._CP14.Currency;
+
+/// <summary>
+/// Allows you to estimate the value of an item with a simple glance.
+/// </summary>
+
+[RegisterComponent]
+public sealed partial class CP14CurrencyExaminableComponent : Component
+{
+}
